refactor: rename segmented button component
