+++ client/src/shared/ui/SegmentedButton/SegmentedButton.tsx
